+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0](https://github.com/aklanti/idempotent/compare/v0.3.0...v1.0.0) - 2026-09-12
+
 ### Breaking changes
 
 - Rename `IdempotencyError` to `Error`, with `EmptyKey`, `KeyTooLong`, `InvalidKey`, `EmptyScope`, and `InvalidScope` in place of `InvalidKey(String)`